clean up/organize utils
